backend/pyproject.toml
[project]
name = "varasaan-api"
version = "0.2.0"
description = "Varasaan backend API"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "aiosqlite>=0.20.0",
  "asyncpg>=0.30.0",
  "boto3>=1.35.0",
  "celery>=5.4.0",
  "email-validator>=2.2.0",
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "passlib[bcrypt]>=1.7.4",
  "pydantic-settings>=2.6.0",
  "pyyaml>=6.0.2",
  "python-jose[cryptography]>=3.3.0",
  "redis>=5.2.0",
  "sentry-sdk>=2.28.0",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.32.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src", "tests"]
norecursedirs = [".tmp", "tmp_pytest", ".pytest_cache"]
markers = ["integration: integration tests"]
cache_dir = "tests/.tmp/.pytest_cache"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
